+++ RELEASENOTES.md
@@ -2,6 +2,9 @@
 
 ### dev-v2 (not yet released) ###
 
+* Decoders: prefer codecs that advertise format support over ones that do not,
+  even if they are listed lower in the `MediaCodecList`.
+
 ### 2.10.0 ###
 
 * Core library: